+++ src/chunk.c
@@ -96,6 +96,7 @@ const char *opcode_nombre(OpCode op) {
         case OP_REPR:            return "OP_REPR";
         case OP_ASIGNAR_REBANADA: return "OP_ASIGNAR_REBANADA";
         case OP_FORMATO_F_SPEC: return "OP_FORMATO_F_SPEC";
+        case OP_LLAMAR_SPREAD_KW_DICT: return "OP_LLAMAR_SPREAD_KW_DICT";
         case OP_LONGITUD:        return "OP_LONGITUD";
         case OP_ES_TUPLA:        return "OP_ES_TUPLA";
         case OP_ES_LISTA:        return "OP_ES_LISTA";