+++ src/debug.c
@@ -260,6 +260,7 @@ int desensamblar_instruccion(const Chunk *c, int offset, FILE *out) {
         case OP_REPR:            return instruccion_simple("OP_REPR", offset, out);
         case OP_ASIGNAR_REBANADA: return instruccion_simple("OP_ASIGNAR_REBANADA", offset, out);
         case OP_FORMATO_F_SPEC: return instruccion_const_corta(c, offset, out);
+        case OP_LLAMAR_SPREAD_KW_DICT: return instruccion_simple("OP_LLAMAR_SPREAD_KW_DICT", offset, out);
         case OP_LONGITUD:        return instruccion_simple("OP_LONGITUD", offset, out);
         case OP_ES_TUPLA:        return instruccion_simple("OP_ES_TUPLA", offset, out);
         case OP_ES_LISTA:        return instruccion_simple("OP_ES_LISTA", offset, out);